Move ParallelFilesProcessor to new services pkg
